Grafana and Prometheus integration -create local setup of grafana and prometheus -added simple dashboard for grafana -added load test -some tests refactoring
