+++ network/local/docker-compose.yaml
@@ -37,10 +37,13 @@ services:
       - ORDERER_GENERAL_TLS_PRIVATEKEY=/etc/hyperledger/crypto/orderer/tls/server.key
       - ORDERER_GENERAL_TLS_CERTIFICATE=/etc/hyperledger/crypto/orderer/tls/server.crt
       - ORDERER_GENERAL_TLS_ROOTCAS=[/etc/hyperledger/crypto/orderer/tls/ca.crt, /etc/hyperledger/crypto/peerOrg1/tls/ca.crt]
+      - ORDERER_OPERATIONS_LISTENADDRESS=0.0.0.0:9444
+      - ORDERER_METRICS_PROVIDER=prometheus
     working_dir: /opt/gopath/src/github.com/hyperledger/fabric/orderers
     command: orderer
     ports:
       - 7050:7050
+      - 9444:9444
     volumes:
         - ../channel-artifacts:/etc/hyperledger/configtx
         - ../crypto-config/ordererOrganizations/luxoft.com/orderers/orderer.luxoft.com/:/etc/hyperledger/crypto/orderer
@@ -55,9 +58,12 @@ services:
       - CORE_PEER_ID=peer0.luxoft.com
       - CORE_PEER_LOCALMSPID=Org1MSP
       - CORE_PEER_ADDRESS=peer0.luxoft.com:7051
+      - CORE_OPERATIONS_LISTENADDRESS=0.0.0.0:9443
+      - CORE_METRICS_PROVIDER=prometheus
     ports:
       - 7051:7051
       - 7053:7053
+      - 9443:9443
     volumes:
         - ../crypto-config/peerOrganizations/luxoft.com/peers/peer0.luxoft.com/:/etc/hyperledger/crypto/peer
     depends_on:
@@ -80,6 +86,29 @@ services:
     depends_on:
       - orderer.luxoft.com
 
-
   ccenv:
     image: hyperledger/fabric-ccenv:1.4.0
+
+  prometheus.luxoft.com:
+    container_name: prometheus.luxoft.com
+    image: prom/prometheus:v2.7.1
+    ports:
+      - 9090:9090/tcp
+    volumes:
+      - ./prometheus.yml:/etc/prometheus/prometheus.yml
+    depends_on:
+      - peer0.luxoft.com
+
+  grafana.luxoft.com:
+    container_name: grafana.luxoft.com
+    image: grafana/grafana:5.4.3
+    environment:
+      - GF_SERVER_ROOT_URL=http://grafana.luxoft.com
+      - GF_SECURITY_ADMIN_PASSWORD=secret
+      - GF_PATHS_PROVISIONING=/etc/grafana/provisioning/
+    ports:
+      - 3000:3000/tcp
+    volumes:
+      - ./grafana/provisioning/:/etc/grafana/provisioning/
+    depends_on:
+      - prometheus.luxoft.com
